Correct algo, use tmp variables to optimize the operations

+++ tests/moving/coordination.h
@@ -1,7 +1,7 @@
 #ifndef COORDINATION_H
 #define COORDINATION_H
 
-void update_main_coordinates(int x_goal, int y_goal, double arrival_angle,
+void update_main_coordinates(int x_goal, int y_goal, float arrival_angle,
                              int r_dep, int r_goal, int n_pts);
 void update_sub_coordinates(void);
